Create README.md
This project was assigned as part of the Orbital Mechanics course during the first year of my Master's program in Space Engineering.

It is composed of two assignments, both of which were carried out using the MATLAB environment to simulate the scenarios and obtain the required results. The first assignment involved analyzing an interplanetary mission and optimizing maneuvers to minimize propellant consumption, while the second focused on studying the effects of recurring perturbations on the orbit of a satellite around Earth.